requirements.txt
# Core
numpy
pandas
scipy
pyyaml
tqdm

# HEP I/O
uproot
awkward
vector

# ML
scikit-learn
xgboost=3.1.2
joblib
shap==0.50.0

# Data formats
pyarrow

# Plotting
matplotlib

# Dev / CI
pytest
ruff

